patient/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
boto3==1.29.7
stripe==7.8.0
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
accelerate==0.25.0
pandas==2.1.4
numpy==1.26.2
mindspore==2.2.0
PyPDF2==3.0.1
python-docx==1.1.0
